# tp-final/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/tenants/{tenantSlug}/servicios": {
      "get": {
        "operationId": "ServiciosController_findAll",
        "parameters": [
          {
            "name": "activo",
            "required": false,
            "in": "query",
            "description": "Filtra por estado. El front del cliente pide `activo=true`; el ABM no manda nada y ve\ntodo. Es un filtro comun y no un flag de autorizacion: un parametro que cambia el nivel\nde acceso de una ruta publica es lo que se rompe en el segundo refactor.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "description": "Cuantos elementos devolver como maximo en esta pagina.",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "El nextCursor que devolvio la pagina anterior. Es opaco: mandalo tal cual.",
            "schema": {
              "maxLength": 512,
              "type": "string"
            }
          },
          {
            "name": "tenantSlug",
            "required": true,
            "in": "path",
            "description": "Identificador del centro en la URL, por ejemplo `lo-de-lili`.",
            "schema": {
              "example": "lo-de-lili",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Una pagina de resultados.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CursorPageDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/ServicioDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "El centro no existe o esta inactivo.\n\nEl centro o el tratamiento no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "summary": "Catalogo de tratamientos del centro, paginado por cursor.",
        "tags": [
          "Servicios"
        ]
      },
      "post": {
        "operationId": "ServiciosController_create",
        "parameters": [
          {
            "name": "tenantSlug",
            "required": true,
            "in": "path",
            "description": "Identificador del centro en la URL, por ejemplo `lo-de-lili`.",
            "schema": {
              "example": "lo-de-lili",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServicioDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServicioDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos invalidos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Falta el token o no es valido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "El token es de otro centro, o de una cuenta que no es de administracion.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "El centro no existe o esta inactivo.\n\nEl centro o el tratamiento no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "Ya hay un tratamiento con ese nombre.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Da de alta un tratamiento. Solo la administradora del centro.",
        "tags": [
          "Servicios"
        ]
      }
    },
    "/api/v1/tenants/{tenantSlug}/servicios/{servicioId}": {
      "get": {
        "operationId": "ServiciosController_findOne",
        "parameters": [
          {
            "name": "servicioId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenantSlug",
            "required": true,
            "in": "path",
            "description": "Identificador del centro en la URL, por ejemplo `lo-de-lili`.",
            "schema": {
              "example": "lo-de-lili",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServicioDto"
                }
              }
            }
          },
          "404": {
            "description": "El centro no existe o esta inactivo.\n\nEl centro o el tratamiento no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "summary": "Un tratamiento puntual.",
        "tags": [
          "Servicios"
        ]
      },
      "patch": {
        "operationId": "ServiciosController_update",
        "parameters": [
          {
            "name": "servicioId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenantSlug",
            "required": true,
            "in": "path",
            "description": "Identificador del centro en la URL, por ejemplo `lo-de-lili`.",
            "schema": {
              "example": "lo-de-lili",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServicioDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServicioDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos invalidos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Falta el token o no es valido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "El token es de otro centro, o de una cuenta que no es de administracion.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "El centro no existe o esta inactivo.\n\nEl centro o el tratamiento no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "Ya hay un tratamiento con ese nombre.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Edita un tratamiento. Es tambien la baja, con `{\"activo\": false}`.",
        "tags": [
          "Servicios"
        ]
      }
    },
...
  },
  "info": {
    "title": "API de turnos",
    "description": "SaaS multi-tenant de turnos para centros de estetica. Cada centro es un tenant y viaja en el path. Los montos son enteros en centavos. Las horas son hora de pared del centro, en formato HH:mm, y las fechas YYYY-MM-DD.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3100",
      "description": "Desarrollo local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ErrorDto": {
        "type": "object",
        "properties": {
          "code": {
            "description": "Identificador de maquina del error. Es el que conviene usar para ramificar.",
            "enum": [
              "validation_error",
              "invalid_cursor",
              "invalid_credentials",
              "unauthenticated",
              "invalid_token",
              "wrong_tenant",
              "tenant_not_found",
              "servicio_not_found",
              "reserva_not_found",
              "servicio_name_taken",
              "ventanas_superpuestas",
              "slot_full",
              "outside_business_hours",
              "past_date",
              "too_far_ahead",
              "invalid_transition",
              "high_contention",
              "forbidden_role",
              "not_found",
              "payload_too_large",
              "too_many_requests",
              "internal_error",
              "http_error",
              "invalid_code",
              "too_many_codes",
              "cliente_not_found",
              "online_payment_unavailable",
              "payment_provider_unavailable",
              "reschedule_not_allowed",
              "too_early_for_no_show",
              "invalid_state",
              "mp_account_change_blocked",
              "mercadopago_not_configured",
              "monthly_limit_reached",
              "plan_limit_reached",
              "mercadopago_not_connected",
              "invalid_signature",
              "slug_taken"
            ],
            "type": "string",
            "example": "slot_full"
          },
          "message": {
            "type": "string",
            "description": "Texto en espanol, listo para mostrarle a una persona.",
            "example": "Ese horario ya no tiene cupo. Elegi otro."
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "CursorPageDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true,
            "description": "Pasalo como `cursor` para pedir la pagina siguiente. null = era la ultima.",
            "example": null
          }
        },
        "required": [
          "data",
          "nextCursor"
        ]
      },
      "ServicioDto": {
        "type": "object",
        "properties": {
          "precioCentavos": {
            "type": "number",
            "description": "Precio total, en centavos. 1800000 son $18.000,00.",
            "example": 1800000
          },
          "senaCentavos": {
            "type": "number",
            "description": "Sena que se cobra al reservar, en centavos. Nunca mayor al precio.",
            "example": 540000
          },
          "reprogramacionHorasAntes": {
            "type": "number",
            "nullable": true,
            "description": "Hasta cuantas horas antes la clienta reprograma sola. null: no puede.",
            "example": 24
          },
          "cancelacionHorasAntes": {
            "type": "number",
            "description": "Hasta cuantas horas antes la clienta cancela con reembolso total.",
            "example": 24
          },
          "id": {
            "type": "string"
          },
          "nombre": {
            "type": "string",
            "description": "Nombre del tratamiento. Unico dentro del centro."
          },
          "descripcion": {
            "type": "string",
            "nullable": true
          },
          "duracionMinutos": {
            "type": "number",
            "description": "Cuanto dura el turno. Define cuantos inicios de grilla entran en cada franja."
          },
          "requiereValoracion": {
            "type": "boolean",
            "description": "Si el tratamiento necesita una valoracion previa con el equipo."
          },
          "activo": {
            "type": "boolean",
            "description": "Un servicio inactivo no se puede reservar. La baja es logica."
          }
        },
        "required": [
          "precioCentavos",
          "senaCentavos",
          "reprogramacionHorasAntes",
          "cancelacionHorasAntes",
          "id",
          "nombre",
          "descripcion",
          "duracionMinutos",
          "requiereValoracion",
          "activo"
        ]
      },
      "CreateServicioDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "description": "Nombre del tratamiento. Unico dentro del centro.",
            "minLength": 2,
            "maxLength": 80
          },
          "descripcion": {
            "type": "string",
            "maxLength": 500
          },
          "duracionMinutos": {
            "type": "number",
            "description": "Duracion del turno en minutos.\n\nEl minimo no es cosmetico: con 0, horaFin queda igual a horaInicio, el predicado de\nsolape no matchea nada y el control de capacidad queda desactivado.",
            "minimum": 5,
            "maximum": 600
          },
          "precioCentavos": {
            "type": "number",
            "description": "Precio total en centavos.",
            "minimum": 0
          },
          "senaCentavos": {
            "type": "number",
            "description": "Sena en centavos. El service valida ademas que no supere al precio.",
            "minimum": 0
          },
          "requiereValoracion": {
            "type": "boolean"
          },
          "reprogramacionHorasAntes": {
            "type": "number",
            "nullable": true,
            "description": "Hasta cuantas horas antes del turno la clienta puede reprogramar sola, sin costo. null\napaga la reprogramacion por autogestion; el centro puede reprogramar siempre.",
            "minimum": 0,
            "maximum": 720
          },
          "cancelacionHorasAntes": {
            "type": "number",
            "description": "Hasta cuantas horas antes la clienta cancela con reembolso total.",
            "minimum": 0,
            "maximum": 720
          }
        },
        "required": [
          "nombre",
          "duracionMinutos",
          "precioCentavos",
          "senaCentavos"
        ]
      },
      "UpdateServicioDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "description": "Nombre del tratamiento. Unico dentro del centro.",
            "minLength": 2,
            "maxLength": 80
          },
          "descripcion": {
            "type": "string",
            "maxLength": 500
          },
          "duracionMinutos": {
            "type": "number",
            "description": "Duracion del turno en minutos.\n\nEl minimo no es cosmetico: con 0, horaFin queda igual a horaInicio, el predicado de\nsolape no matchea nada y el control de capacidad queda desactivado.",
            "minimum": 5,
            "maximum": 600
          },
          "precioCentavos": {
            "type": "number",
            "description": "Precio total en centavos.",
            "minimum": 0
          },
          "senaCentavos": {
            "type": "number",
            "description": "Sena en centavos. El service valida ademas que no supere al precio.",
            "minimum": 0
          },
          "requiereValoracion": {
            "type": "boolean"
          },
          "reprogramacionHorasAntes": {
            "type": "number",
            "nullable": true,
            "description": "Hasta cuantas horas antes del turno la clienta puede reprogramar sola, sin costo. null\napaga la reprogramacion por autogestion; el centro puede reprogramar siempre.",
            "minimum": 0,
            "maximum": 720
          },
          "cancelacionHorasAntes": {
            "type": "number",
            "description": "Hasta cuantas horas antes la clienta cancela con reembolso total.",
            "minimum": 0,
            "maximum": 720
          },
          "activo": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}